fix index and update

+++ app/Http/Controllers/ProjectController.php
@@ -4,9 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Project as ModelsProject;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
-use SebastianBergmann\CodeCoverage\Report\Xml\Project;
 
 class ProjectController extends Controller
 {
@@ -17,8 +17,10 @@ class ProjectController extends Controller
      */
     public function index()
     {
+        $projects = Auth::user()->projects()->orderBy('id', 'desc')->paginate(25);
+        // dd($projects);//mmmyyy
         return view('projects.index', [
-            'projects' => 'projects',
+            'projects' => $projects,
         ]);
     }
 
@@ -53,23 +55,27 @@ class ProjectController extends Controller
         $project->start_date = $request->startDate;
         $project->stop_date = $request->stopDate;
         $project->message = $request->message;
-        $project->image = $this->uploadImage('image', $request);
         $project->user_id = auth()->user()->id;
-        try {
-            DB::beginTransaction();
+
+        if($file = $request->file('image')){
+            $project->image = $this->uploadImage($file, $request);
+            try {
+                DB::beginTransaction();
+                $project->save();
+                DB::commit();
+            } catch (\Exception $e) {
+                DB::rollback();
+                Storage::delete('public/' . $project->image);
+                throw $e;
+            }
+        }else{
             $project->save();
-            DB::commit();
-        } catch (\Exception $e) {
-            DB::rollback();
-            Storage::delete('public/' . $project->image);
-            throw $e;
         }
         return redirect()->route('projects.index');
     }
 
-    protected function uploadImage($name, $request)
+    protected function uploadImage($file, $request)
     {
-        $file = $request->file($name);
         $fileName = $file->getClientOriginalName();
         $name = pathinfo($fileName, PATHINFO_FILENAME) . '-' . $file->hashName();
         $path = 'storage/' . $file->storeAs(
@@ -99,7 +105,11 @@ class ProjectController extends Controller
      */
     public function edit($id)
     {
-        //
+        $project = ModelsProject::findOrFail($id);
+        $this->authorize('update', $project);
+        return view('projects.edit', [
+            'project' => $project,
+        ]);
     }
 
     /**
@@ -111,7 +121,34 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $request->validate([
+            'name' => 'required|unique:projects|max:255',
+            'startDate' => 'required',
+            'stopDate' => 'required',
+            'message' => 'nullable',
+            'image' => 'nullable|file|mimes:jpg,jpeg,png|max:2048',
+        ]);
+        $project = ModelsProject::findOrFail($id);
+        $this->authorize('update', $project);
+        $project->name = $request->name;
+        $project->start_date = $request->startDate;
+        $project->stop_date = $request->stopDate;
+        $project->message = $request->message;
+        if($file = $request->file('image')){
+            $project->image = $this->uploadImage($file, $request);
+            try {
+                DB::beginTransaction();
+                $project->save();
+                DB::commit();
+            } catch (\Exception $e) {
+                DB::rollback();
+                Storage::delete('public/' . $project->image);
+                throw $e;
+            }
+        }else{
+            $project->save();
+        }
+        return redirect()->route('projects.index');
     }
 
     /**
